Log an exception rather than erroring out when data is found for metrics that are not attached to the experiment (#2680)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2680

D57165887 added an error when modeling layer encountered data for metrics that are not attached to the experiment. Daniel has been seeing these errors in some use cases where it is non-trivial to attach these metrics to the experiment. In order to unblock, we are relaxing the error, and logging an exception instead. This should keep the issue visible while letting the experiment proceed.

Reviewed By: danielcohenlive

Differential Revision: D61544922

fbshipit-source-id: dd843bc205ac61c5039e6352c064f16bc077765b