Change time limit for unit tests to 60 seconds and introduce a decorator for long running tests (#2661)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2661

This diff is to change the time limit for unit tests to 60 seconds and introduce a decorator for long running tests.

We'll use the `ax_long_text` decorator to identify all long running tests, which we can fix over time by creating backlog tasks for test-owners to update these long-running tests with a descriptive message.

This change also introduces this decorator to all tests running over 60 seconds. For now, the provided reason for these long running tests are placeholders.

Reviewed By: mpolson64

Differential Revision: D61212556

fbshipit-source-id: d6c3e343f71d472e9e0ea8d151cd07c21ca724c8