[Website] Use mathjax svg renderer
We identified an issue where mathjax would fail to render certain math symbols. This happens on MacOS versions Ventura and later (13+) because of a change in the OS-supplied STIX fonts. We fix this here by changing the renderer from HTML+CSS (which tries to use local STIX fonts) to SVG.

This is a follow up to https://github.com/pytorch/botorch/pull/2481 which solved the same problem on the botorch website