Make sure is_ax_equal returns a boolean
Summary:
This is an alternative to D61666276 that addresses some of the issues raised there.
Instead of checking for type equality, which is too restrictive, we attempt to cast the output of the equality comparison to a boolean. If the cast or the comparison fails, we return False. This ensures that the output of `is_ax_equal` is always a boolean and can be used in `if` statements.

Also documented some of the special cases used in this method, as its behavior can sometimes be unintuitive.

Differential Revision: D61803910