Make sure is_ax_equal returns a boolean (#2714)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2714

This is an alternative to D61666276 that addresses some of the issues raised there.
Instead of checking for type equality, which is too restrictive, we attempt to cast the output of the equality comparison to a boolean. If the cast or the comparison fails, we return False. This ensures that the output of `is_ax_equal` is always a boolean and can be used in `if` statements.

Also documented some of the special cases used in this method, as its behavior can sometimes be unintuitive.

Reviewed By: esantorella

Differential Revision: D61803910

fbshipit-source-id: e555e3feaee2802b3a6651e6eb8e3e5d6257992c