Implement Data/MapData.clone and use it in Experiment.clone_with (#2703)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2703

Prior to this change, the data attached to the cloned experiment would retain the `_db_id` of the original. When the clone experiment is saved, this would lead to overwriting the DB entry to point to the clone experiment, effectively deleting the data attached to the original experiment.

This diff implements `clone` methods for `Data` and `MapData`, and uses it in `Experiment.clone_with` to avoid this issue.

Reviewed By: ItsMrLin

Differential Revision: D61715842

fbshipit-source-id: fd849bea33af93a10245f4662f4b3c4886e353ad