Get rid of MultiObjectiveBenchmarkProblem (#2721)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2721

Context:

The only purpose of `MultiObjectiveBenchmarkProblem` is as a type annotation indicating whether its `optimization_config` is a `MultiObjectiveOptimizationConfig`. This creates more trouble than its worth; we can simply check whether the optimization config is multi-objective rather than checking whether the problem is multi-objective. The same goes for having single-objective and multi-objective surrogate problems.

With this change, the only two `BenchmarkProblem` classes are `BenchmarkProblem` and `SurrogateBenchmarkProblem`; the latter can be removed in a future PR.
This diff:
* Removes `MultiObjectiveBenchmarkProblem`, and replaces references to it with `BenchmarkProblem`
* Consolidates `SurrogateBenchmarkProblemBase`, `SooSurrogateBenchmarkProblem`, and `MOOSurrogateBenchmarkProblem` into `SurrogateBenchmarkProblem` and replaces references
* Removes some branching logic

Reviewed By: Balandat

Differential Revision: D61869477

fbshipit-source-id: f5a7948d264616bf4a8e7352cf851d68ce303dc6