Evaluate oracle at target task and fidelity (#2727)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2727

`BenchmarkRunner.evaluate_oracle` now evaluates at the target task and fidelity.

A weakness of this solution is that it requires passing the search space to the runner. Generally, there is too much information that must be on both the problem and the runner; this could be addressed by making the problem an attribute of the runner rather than the other way around.

Reviewed By: sdaulton

Differential Revision: D61884295

fbshipit-source-id: 76ae3f982ad02ea4e999f66c27a0f28b90a3ff00