Replace Models.GPEI usage with Models.BoTorchModular
Summary: `Models.GPEI` points to the deprecated legacy Ax model. For most practical usage, it behaves identical to `Models.BOTORCH_MODULAR` (aka MBM), which is the recommmended model to use in Ax. This diff updates many use cases to make it easier to fully deprecate the registry entry down the line.

Differential Revision: D62146528