Replace Models.GPEI usage with Models.BoTorchModular (#2733)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2733

`Models.GPEI` points to the deprecated legacy Ax model. For most practical usage, it behaves identical to `Models.BOTORCH_MODULAR` (aka MBM), which is the recommmended model to use in Ax. This diff updates many use cases to make it easier to fully deprecate the registry entry down the line.

Reviewed By: Balandat

Differential Revision: D62146528

fbshipit-source-id: 7f53f9748432b70c477690d20959e21809122435