Actions: Publish Website use github actions bot
The previous code attempted to create these same strings using the actions `github` environment variable but these are not available in the shell script. It may be possible to pass in the variable but we can just be explicit here instead that this is intended to be done by the bot.

The name and email are provided by the official actions/checkout@v4 documentation: https://github.com/actions/checkout?tab=readme-ov-file#push-a-commit-using-the-built-in-token