Move _get_model_state_from_last_generator_run from GS to GNode (#2738)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2738

Background: `_get_model_state_from_last_generator_run` is used to extract the model state from last GR (generated by the current model), which is then used to ensure the current model is being fit with the necessary model state.

The chicken & egg problem (see T199011652): In order to extract the correct model state, we need to know what the current model is. If a `GNode` has multiple model specs, we need to fit both and do model selection to determine what the current model is. But we need to extract the state before we can fit the models, so this logic doesn't work at the `GS` level. It has to be done separately for each `ModelSpec` being fit.

This diff:
- Replaces `GS._get_model_state_from_last_generator_run` with `GNode._get_model_state_from_last_generator_run`.
- The new method operates on `ModelSpec` level, extracting the state only from previous GRs of the same model.

Design considerations: Opted for doing this on `GNode` rather than on `ModelSpec`, since we can access the previous GRs on `GNode` through its `GS` backpointer. If we were to move this down to `ModelSpec`, we'd have to pass the GRs down, which I didn't want to do.

Reviewed By: lena-kashtelyan

Differential Revision: D62136042

fbshipit-source-id: 1f0ce1baee2400242744be9cd494f37d76dd5744