Docusaurus: Automated migration of md and pages
Using `npx @docusaurus/migrate migrate --mdx --page website website-v2`

The key here is that we included the `--mdx` and `--page` flags to the migration tool. The main change here is that docusaurus has migrated from MD to MDX which is more akin to JSX and requires some adjustments to our markdown docs. The tool is heavy handed and fixes most issues but there will be manual formatting changes and migrations coming in future commits.