Bug fix: Make `Acquisition.optimize` work with discrete optimizer regardless of whether `raw_samples` is in `optimizer_options` (#2772)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2772

Context: When `raw_samples` is present in `optimizer_options` and the discrete optimizer is used, `raw_samples` is removed from `optimizer_options` because it isn't appropriate for for the discrete optimizer. However, `raw_samples` can't be removed if it isn't present. Not passing it (which is better) causes an error.

This PR: Doesn't try to remove `raw_samples` when it isn't present.

Reviewed By: saitcakmak

Differential Revision: D63035021

fbshipit-source-id: 7ad947ede4e944ff2dfea816afd67c88e3f3f496