Remove hypothesis dependency (#2774)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2774

Context: Hypothesis is not commonly used in Ax and is not part of the `UNITTEST_MINIMAL_REQUIRES` dependencies (although it is part of `UNITTEST_REQUIRES`). It is used in only one benchmarking test. Removing it will make it easier to build an environment that can run benchmark tests. We could have a broader discussion about removing it from Ax entirely.

This diff: Changes a unit test to not use hypothesis

Reviewed By: saitcakmak

Differential Revision: D63150787

fbshipit-source-id: daf732bb9effc45406816610f0d95f547d5ab192