Use input constructors during GS
Summary:
This diff adds support for leveraging the input constructors during gs.gen_with_multiple_nodes(..), as part of this we also add _previous_node param on gennodes. We presist the arms_per_node argument as we believe this is likely useful for researchers and developers internally, but this should never be exposed at api level to not cause confusion.

we also add a e2e style unit test

Follow up diffs:
- storage --> let's do this once we all like the 3 input constructors
- update the input constructors to handle the case where n isn't provided as a kwarg
- add tests for real world scenario per Liz suggestion
- add test to enforce signature across input constructors per Liz/Sait/Daniel suggestion

Differential Revision: D62316641