Allow removal of parameters with different values in RemoveFixed transform (#2779)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2779

Previously, the `RemoveFixed.transform_observation_features` transform checked that the parameter being removed had the same value as the fixed parameter in the search space used to initialize the transform.
This diff removes this check to allow removal of any observation parameter with the same name. This will allow the transform to operate on observations from two similar but non-identical search spaces.

Reviewed By: susanxia1006

Differential Revision: D63322396

fbshipit-source-id: 2648a1cac171a871229101fcc1d44c4200f58d1d