Parameterization type fixes; use parameters instad of arms where appropriate (#2780)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2780

Sometimes benchmarking code creates Arms from parameters just so they can be passed to a function that expects an Arm but only uses its parameters. This is silly. It's better to just have the function expect parameters.

Also updated some method signatures to use `Mapping` to indicate that they do not mutate the parameterization, which unfortunately creates the need for Pyre-fixmes if they are passed to a function or class such as `Arm` that does not annotate its arguments as immutable.

Reviewed By: Balandat

Differential Revision: D63327381