Implement an __init__ method for GenerationStep
Summary:
Dataclasses by default generate an `__init__` method based on the fields specified in dataclass definition. This is generally quite convenient but it leads to some issues in the case of `GenerationStep`.
Some of the inputs to `GenerationStep` constructor is used to construct the underlying `ModelSpec`. With the default constructor, certain attributes that we would expect to be identical, such as `GenerationStep.model_kwargs` and `ModelSpec.model_kwargs`, can go out of sync if one of them is updated after initialization.
Because dataclass fields are class attributes (including `InitVar`s), we cannot define properties with the same name linking the attributes directly to the attributes of the underlying `ModelSpec`. Implementing a custom `__init__` method lifts this restriction.

Note that I didn't add setters for `model_kwargs` and `model_gen_kwargs`. I found only one place in the codebase where they were being updated, so it didn't seem necessary.

Differential Revision: D63497022