Predicted Effects Plot (#2777)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2777

This replicated the functionality of `ax.plot.scatter.plot_fitted`.  It will show predicted effects for all trials with data and the most recently created non abandoned trial (which may have data also).  The intent is to use it for a candidate trial which has just been generated, which will be the case in scheduler.  But in the event the most recent trial is not a candidate, it should still work.

To come:
- setting trial index
- mark arms that violate constraints
- Sane Limit for number of in-sample arms?

Reviewed By: mpolson64

Differential Revision: D62325402