Simplify passing 'use_model_predictions' for best-point selection (#2804)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2804

Context: Currently, a `BenchmarkMethod` has `best_point_kwargs` that are passed to `BestPointMixin`. When passing kwargs like this, it's hard to make sure defaults are correct, and in fact MBM benchmarks were inadvertently using model predictions because they were passing `best_point_kwargs={}`, overriding the default that was filled in when best_point_kwargs are not passed at all.

This is unnecessarily complex, because the only argument currently supported in `best_point_kwargs` is `use_model_predictions`.

This PR:
* Changes passing `best_point_kwargs` to just passing a boolean

Reviewed By: saitcakmak

Differential Revision: D63653755