Add _ipython_display_ to AnalysisCards (#2803)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2803

Adding _ipython_display_ allows display(...) to inteligently consume and output the cards in a way that is beneficial for the user.

By overriding this method in each of our Card subclasses we can call the the same display(...) method and the notebook will correctly select whether to show the dataframe, plotly plot, or markdown text.

Note that passing diplay a list of cards will not result in display being called for each element, so I've also added a display_cards method that iterates through the list for the user.

Reviewed By: mgarrard

Differential Revision: D63649339