Tutorials: Update styling for output cells and plots
Docusaurus introduced a handy table of contents on the right side of docs pages which causes the available space in my window to be less than the hardcoded width for plotly diagrams (hardcoded by us in the Ax library). I handle that here by adding a scrollbar to the plot cell when necessary.

Also make some tweaks to the styling of output cells to make them respect dark mode and be more compact.