Make parameters ordered in SearchSpace
Summary:
This PR: Changes `SearchSpace.parameters` from a `dict` to an `OrderedDict`.

Context: In benchmarking, the order of parameters matters because parameters can be converted to a tensor. In the current Python version, this should work even without using `OrderedDict` because dicts are insertion-ordered, but this isn't an officially supported language feature or guaranteed to last.

One downside is that by annotating the parameters as an `OrderedDict`, we can't also annotate them as an immutable `Mapping`. For now, this is fine, because the parameters are mutated, so we can't annotate them as immutable anyway.

Differential Revision: D63641392