Tutorials: Tranform style attributes *after* running mdformat
Rendering errors surfaced a bug in the script where mdformat doesn't like our JSX if it contains the transformed style attributes. It would think the "html" is invalid and add a backslash to i guess try and escape it?

This was a confusing bug, thankfully someone else ran into this before and offered some guidance: https://github.com/executablebooks/mdformat/issues/388#issuecomment-1483371326