Fix Experiment.clone_to data mapping
Summary:
`Experiment.clone_to` accepts a `trial_indices` argument, which specifies the list of trial indices to keep on the expeirment. Prior to this change, the data from the old trials was being cloned and attached to the clone experiment without modification. This was leading to the old trial index being retained on the attached data.

This diff updates the trial index before attaching the data and makes sure to attach the data to the correct index for the cloned trial.

Differential Revision: D63725668