Fix Experiment.clone_to data mapping (#2807)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2807

`Experiment.clone_to` accepts a `trial_indices` argument, which specifies the list of trial indices to keep on the expeirment. Prior to this change, the data from the old trials was being cloned and attached to the clone experiment without modification. This was leading to the old trial index being retained on the attached data.

This diff updates the trial index before attaching the data and makes sure to attach the data to the correct index for the cloned trial.

Reviewed By: Balandat

Differential Revision: D63725668

fbshipit-source-id: 2a758bab94031c51a080a06c0d5f3adc430a0790