Always predict for status quo features trial index (#2810)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2810

Previously we had plots that looked like {F1904605463}
After the same plot looks like
{F1904608806}

##What's happening
Because of nonstationarity, the same feature gets predicted differently in different trials.  But the point of this plot is to show how well the parameterization does on the metric, so we want to standardize the trial index we're predicting on.

Differential Revision: D63715134

Reviewed By: mgarrard