Validate the inputs to ModelBridge.predict
Summary:
This came out of a puzzling bug in a notebook that took some time to debug. If you pass in a list of `Arm`s, `ModelBridge.predict` runs (since `Arm` also has `parameters`) but it doesn't apply transforms to the parameters (since `Arm.parameters` returns a clone of `Arm._parameters`). This can lead to non-sense predictions without the user realizing there is an issue with their code (in this case an `OverflowError:math range error` flagged the issue to us).

This diff adds a simple validation in `ModelBridge.predict` to check that the input is actually a list of `ObservationFeatures`.

Differential Revision: D64115012