Make benchmark runners into dataclasses (#2846)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2846

Context: Using dataclasses makes things easier , and will make things *much* easier if we eventually rework serialization for BenchmarkRunners to not rely on SerializationMixin (as in D64061291). It also allows for stricter equality checks in a natural way.

This PR:
* Makes `BenchmarkRunner`, `SyntheticProblemRunner`, into dataclasses, without changing their init signatures or making substantive changes to their attributes.
* Renames attributes with underscores like `_test_problem_class` to `test_problem_class`, so that the attributes match what is used for initialization, to make these work as dataclasses
* Introduces a stricter equality check that uses `dataclasses.asdict`

Note: I had hoped to make this into frozen dataclasses, but this doesn't combine well with setting attributes in the post-init. That can be done but ia little opaque; see even though this is not really mutating the object after initialization, see https://stackoverflow.com/questions/53756788/how-to-set-the-value-of-dataclass-field-in-post-init-when-frozen-true

Reviewed By: Balandat

Differential Revision: D64110932

fbshipit-source-id: 63b509de8e08d93873250a6adb5f3834bbdabb4a