Update datetime equality checks to use a tolerance of 1 second
Summary: These previously ignored microseconds and compared equality up to the second. This would lead to strange failures when the timestamps differed by only a fraction of a second but one crossed over to the next second. Simply using a tolerance of one second should simplify things and reduce flaky failures.

Differential Revision: D64218482