Update datetime equality checks to use a tolerance of 1 second (#2857)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2857

These previously ignored microseconds and compared equality up to the second. This would lead to strange failures when the timestamps differed by only a fraction of a second but one crossed over to the next second. Simply using a tolerance of one second should simplify things and reduce flaky failures.

Reviewed By: mgarrard

Differential Revision: D64218482

fbshipit-source-id: 0e0eea10411c18eb5e701813db8fad68a3cd3313