Fix is_optimal sometimes being used for y_metric_name
Summary:
When generating the plot the y_metric name is inferred based on the other columns expected to be. When is_optimal was added this needed to be updated as well.

Now, instead of inferring we're just going to supply the y metric name directly -- this should be less error prone.

Reviewed By: mgarrard

Differential Revision: D64123183

fbshipit-source-id: 20186b1964aee561ebc81c484cb8b4dc5125efbf