Add experiment to input constructor signature (#2871)
Summary:

In a previous diff both Daniel and Lena suggested adding experiment to the method signatures for input constructors, at the time we only had a single input constructor that needed the exp object, but in coming diffs we'll grab it for other input constructors. If we continue to rely on gs_kwargs to get it, we have to do null checks etc, and this is much better.

Reviewed By: danielcohenlive, lena-kashtelyan

Differential Revision: D64308015