Use input constructors to determine fixed features to pass to model (#2877)
Summary:

In previous diffs we added the logic to input constructors to support identification of the target trial, which is a fixed feature, based on an algorithm. In this diff we include grabbing fixed features during generation in both ```gen_for_multiple_trials_with_multiple_nodes``` and ``gen_with_multiple_nodes```. If a user passes in fixed features those will take precedence over algorithmically determined values. We also don't smush together passed in fixed features as we don't have a consistent way of distinguishing target_trial fixed features from other fixed features. For this reason, if a user passes in fixed features it's important that they actually specify **all** fixed features.

The unit tests I add here don't actually use the fixed_features in modeling but ensure that they are passed to the model gen method. In follow up diffs we will:
1. use target trial in dispatch logic
2. expose target trial override in make_candidate_trials
3. add e2e test showing that with this input constructor we can correctly gen from multi-task mbm

Reviewed By: danielcohenlive

Differential Revision: D64288411