Add logs for when data-fetching actually starts (#2855)
Summary:

Current logs are confusing, they make it look like we first fetch the data quite fast, then we retrieve completed trials, and then we do nothing for a while. This clarifies by rephrasing the "Fetching..." log to "will fetch" and logging the actual start of fetching with "..." at the end, indicating that it will take a bit.

Reviewed By: Balandat

Differential Revision: D64208425