Improve signature on is_met() method on transition criterion (#2902)
Summary:

In a previous diff in this stack (52) we added an argument to pass the current node into tc.is_met(). This is actually great news because now we don't need a lot of these other optional parameters!

This diff cleans up those optional parameters, and replaces their usage by directly pulling it from the node object.

We also fix all calls to tc.is_met() to use the new signature

Differential Revision: D64491588