Fix pending point extraction in gen_multi_from_multi (#2914)
Summary:

Earlier today danielcohenlive was having trouble using gs and was getting a weird modeling error. Lena did some investigation an found we were using the wrong method to grab pending points from the experiment. See her notebook in the test plan for more details. to quote "Basically in some settings it’s not safe to deem a point pending based on the trial status; generally no point for which we have data, we want to consider pending"

I just am putting up the fix with her logic here

Reviewed By: lena-kashtelyan, Cesar-Cardoso

Differential Revision: D64563496