Remove support(?) for multiple surrogates from Acquisition (#2916)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2916

The current implementation of `Acqusition.__init__` technically supports multiple surrogates. It combines the models from each surrogate into a `ModelDict` before passing them down to acquisition input constructors. `ModelDict` was added some time ago with the goal of using it to support failure aware BO, but those methods never got implemented and there is no current use case that supports it.

In its current form, the support for multiple surrugates is completely superficial and does not bring much value. It requires complex handling of arguments like `X_pending` and `X_observed`, which itself comes with TODOs to fix or improve it.

This diff removes support for multiple surrogates from `Acquisition` and cleans up some of the complicated argument handling that was necessitated by it. If we decide to support multiple surrogates again at a later date, we can do so with a better thought out design and implement it more cleanly.

Reviewed By: lena-kashtelyan

Differential Revision: D64610244

fbshipit-source-id: 70479e4ca7a5c7fc99ea36ed7c95261d1a2fd4a9