Remove unneeded `optimal_value` and `_is_constrained` from `ParamBasedTestProblem` and `_is_constrained` and `_is_moo` from `SyntheticProblemRunner`
Summary:
Context: Several parameters have become redundant during refactoring. Also, a `ParamBasedTestProblem` should be as minimal as possible, as the only benchmark class that needs to be repeatedly subclassed.

`optimal_value` does not need to be defined on a `ParamBasedTestProblem`, since it is also defined on the `Problem`; logically, it should be with the problem, because the optimal value depends on the search space and optimization config. A `ParamBasedTestProblem` should only generate data.

`_is_moo` and `_is_constrained` are helper functions that are not very helpful, with `_is_constrained` being redundant.

This diff:
* Removes `optimal_value` from `ParamBasedTestProblem`, which is defined redundantly on the `BenchmarkProblem`
* Removes `_is_moo` from `SyntheticRunner`, which only has one usage which isn't very helpful
* Removes `_is_constrained` from `ParamBasedTestProblem`, which is defined redundantly on `SyntheticRunner`
* Removes custom equality check from `ParamBasedTestProblem`, so it can just use dataclass equality method.
* Makes some minor fixes
* Makes the equality method for `ParamBasedTestProblem` more rigorous by removing it, so it uses the dataclass equality check

Differential Revision: D64572527