Abstract away a `FuncEnum` (will use it in `best_model_selector.ReductionCriterion` and `NodeInputConstructors`) (#2887)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2887

Add a base class for any "enums of callables" that will actually play well with storage and provide a consistent and safe way to implement such enums.

NOTE: Once we agree on the implementation, I'll add storage and e2e tests that leverage this in GenerationStrategies with model selection.

Reviewed By: mgarrard

Differential Revision: D63914989

fbshipit-source-id: bcdf59330985b243c9e2a20fa6cf612619ee20a1