Isolate the "no feasible points" error to objective threshold inference during MOO (#2929)
Summary:

### The problem

For a MMO problem, if no thresholds are provided on the objectives, we try to infer them during gen().  If there are outcome constraints, and none of the observed data points are feasible, we cannot infer the thresholds and error out. This is reproduced in N5821958.

However, the error makes it look like candidate generation failed, when in fact it is the inference of the thresholds.  We could simply suggest the users to provide thresholds if they can and the error will be gone.

### The caveats

The same error message are being referenced in multiple places.
In particular, it is being used in PosteriorMeanPriorSampler, and acquisition functions for non-MBM models.  However, in those cases, we do not actually filter for feasibility/outcome constraints.  Instead, we only filter for points where all metrics (objectives and outcome constraints) are attached, and sometimes parameter constraints as well.  So in such cases, the error should happen very rarely, and the error message is misleading at best.

### This diff

Change the error message to a more actionable when incurred during reference point inference.  And change the error from all other places mentioned above to "no observed points" error message.

Reviewed By: saitcakmak

Differential Revision: D64691734