Move noise_std, constraint_noise_std, and negate from `ParamBasedTestProblem` to `SyntheticProblemRunner` (#2926)
Summary:

Context: Noise and negation are (somewhat surprisingly!) handled by the runner, even though similar attributes exist on BoTorch test problems. This is confusing; it is better to _require_ these to be set on the runner and to raise an exception if they are set on the test problem.

Also, a `ParamBasedTestProblem` should be as minimal as possible, since it is the only benchmark class that needs to be repeatedly subclassed.

Furthermore, this makes the code easier to work with by moving these arguments to a shallower level in the stack, making them easier to access and reducing the need to pass dicts of parameters.

This PR:
* Adds `noise_std`, `constraint_noise_std`, and `negate` to `SyntheticProblemRunner`
* Removes those arguments from `ParamBasedTestProblem` and all its subclasses
* Updates references
* Adds an exception if those arguments are present when creating a runner based on a BoTorch problem

Reviewed By: Balandat

Differential Revision: D64575398