Add test of correct result for SEBO (#2943)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2943

Adds a unit test  that SEBO produces the correct result in a
controlled test case. Changes the existing test case data to match the desired
test case for this test, which is a linear function with two unused variables.
We verify that we maximize the used variable, while setting the unused
variables to their target point values.

Depends on D60089765

Reviewed By: dme65

Differential Revision: D64720328

fbshipit-source-id: 98a4b078300bfcc47ca60bc65bfc6c470385738f