Redo constraint violation indicators and legend on insample and predicted plots (#2955)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2955

Instead of varying sizing, we're adding a red outline of varying opacity based on the overall probability of constraint violation.

There is a hack to make sure the outlines don't appear in the legend, as the legend normally reflects the first point in each group.  We're hiding the original and creating a duplicate group of the same name at point (None, None).  In doing this, we give up the ability to toggle points on and off the graph by clicking on their legend item, which I just discovered existed.

{F1944580069}

Reviewed By: ItsMrLin

Differential Revision: D64850289

fbshipit-source-id: af1eb16f0c19b24f68e51417791b789f9d0fa476