Update TrialAsTask transform to default pending points to most recent trial (#2965)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2965

For MT mbm implementations, we use the TrialAsTask transform, however, until now this has only been used in scenarios where the trial is composed from a singular model. If we want to extend trial composition to be from multiple models, if the first model that is called is to create the trial generates a set of points, those become pending points. During fit and gen, the trialastask transform is called, and when we go to transform the pending points from that model within the current trial, it doesn't have an associated trial, so then we'll end up with a ```KeyError: 'TRIAL_PARAM' ``` error.

Reviewed By: sdaulton

Differential Revision: D64974517

fbshipit-source-id: 0ae13d9659c7aa976c4884210c5fc580f20a555c