Fix test_init_files (#2982)
Summary:

I randomly came across this test that would've prevented a recent broken release if it has been working. I believe it was written based on the old directory structure (before we moved ax/ax/... to ax/...), and had to be updated to continue working. This diff updates the target directory and adds exclusions for
- ax/fb -- these are only used with buck, so no need to enforce __init__
- ax/github -- these are for github templates and workflows, so no need to have them included in builds. They also live in a separate directory in OSS.
- tests -- I don't have a strong opinion on this one either way

Reviewed By: esantorella

Differential Revision: D65141273