Add `optimize_acqf_mixed_alternating` to `mock_botorch_optimize_context_manager` & reduce duplication with `mock_optimize_context_manager` (#2973)
Summary:

A previous diff added mixed optimizer to MBM. This diff adds it to optimizer mocks.

`mock_botorch_optimize_context_manager` had a good bit of overlap with BoTorch's `mock_optimize_context_manager`, which is also cleaned up in this diff. It now uses `mock_optimize_context_manager` and adds additional mocks on top of that.

Differential Revision: D65067691