Add `optimize_acqf_mixed_alternating` to `mock_botorch_optimize_context_manager` & reduce duplication with `mock_optimize_context_manager` (#2973)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2973

A previous diff added mixed optimizer to MBM. This diff adds it to optimizer mocks.

`mock_botorch_optimize_context_manager` had a good bit of overlap with BoTorch's `mock_optimize_context_manager`, which is also cleaned up in this diff. It now uses `mock_optimize_context_manager` and adds additional mocks on top of that.

Reviewed By: paschai

Differential Revision: D65067691

fbshipit-source-id: 47185e63e6e462c843d55f29d031be35583d8b05