Use input constructors to define status quo features for now (#2990)
Summary:

In the future we'd like to pursue an option outlined by sdaulton: "we would still need to set target trial in the transform_config for Derelativize in model_kwargs, unless we added some logic like use latest trial or something in derelativize. It seems like target trial is definitely the trial index that we should use"

But we need to be able to support relative constraints in next mondays launch so this seemed easier and faster for me to implement. Would like any thoughts on if this is an acceptable workaround for the next 2-4weeks?

Reviewed By: sdaulton

Differential Revision: D65161089